backend/requirements.txt
fastapi
uvicorn
python-dotenv
supabase
openai
PyMuPDF
python-multipart
pytesseract
pdf2image
pdfplumber